Factor out ParseErrorListener and use in Preproc
