+++ src/main/scala/alogic/AParser.scala
@@ -1,34 +1,9 @@
 package alogic
 
-import org.antlr.v4.runtime.{ ANTLRInputStream, CommonTokenStream }
+import org.antlr.v4.runtime.ANTLRInputStream
+import org.antlr.v4.runtime.CommonTokenStream
 
-// TODO figure out how to do preprocessor
-//   A) Put into VParser.g4
-//       Problem is that there are many places where a #if could appear, controlling ports/modules/statements and the grammar seems to explode
-//   B) Put into VLexer.g4 using actions
-//     This feels like the right place, but implementation is unclear.
-//       Problem is that it is not clear whether we can use the same trick of parsing header file once, and then reusing the defines
-//       Perhaps can make lexer once and give it more input?
-//       Perhaps can add method to adjust the defines dictionary?
-
-import scala.io.Source
-import java.io.File
-import org.antlr.v4.runtime.RecognitionException
-import org.antlr.v4.runtime.Recognizer
-import org.antlr.v4.runtime.BaseErrorListener
 import scalax.file.Path
-
-object ParserErrorListener extends BaseErrorListener {
-  override def syntaxError(recognizer: Recognizer[_, _],
-                           offendingSymbol: Object,
-                           line: Int,
-                           charPositionInLine: Int,
-                           msg: String,
-                           e: RecognitionException) = {
-    val loc = Loc(recognizer.getInputStream.getSourceName, line)
-    Message.error(loc, s"Syntax error: $msg")
-  }
-}
 
 class AParser(includeSearchPaths: List[Path]) {
