+++ src/main/scala/alogic/ParseErrorListener.scala
@@ -1,0 +1,17 @@
+package alogic
+
+import org.antlr.v4.runtime.RecognitionException
+import org.antlr.v4.runtime.Recognizer
+import org.antlr.v4.runtime.BaseErrorListener
+
+object ParserErrorListener extends BaseErrorListener {
+  override def syntaxError(recognizer: Recognizer[_, _],
+                           offendingSymbol: Object,
+                           line: Int,
+                           charPositionInLine: Int,
+                           msg: String,
+                           e: RecognitionException) = {
+    val loc = Loc(recognizer.getInputStream.getSourceName, line)
+    Message.error(loc, s"Syntax error: $msg")
+  }
+}
